Time: 695 ms (42.38%), Space: 208.4 MB (28.97%) - LeetHub

+++ 2316-count-unreachable-pairs-of-nodes-in-an-undirected-graph/2316-count-unreachable-pairs-of-nodes-in-an-undirected-graph.cpp
@@ -25,7 +25,7 @@ public:
             graph[v].push_back(u);
         }
         
-        vector<int> comp;
+        vector<long long> comp;
         
         for(int i = 0; i < n; i++)
         {
@@ -34,7 +34,7 @@ public:
         }
         
         long long res = 0;
-        long long cnt = n;
+        int cnt = n;
         
         for(auto nodes : comp)
         {
